requirements.txt
importlib-metadata >= 1.0 ; python_version < "3.8"
progressbar2 >= 4.2.0
